fixing bug introduced by bad widget keys
